Deleted MainActivity2 and MainActivity3
