--- app/src/main/java/com/example/olympia/MainActivity3.java
@@ -1,14 +1,0 @@
-package com.example.olympia;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class MainActivity3 extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main3);
-    }
-}
