+++ app/src/main/res/layout/activity_main2.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/useme"
-    tools:context=".MainActivity2">
+    tools:context=".MainActivity">
 
     <TextView
         android:id="@+id/signin"
