+++ app/src/main/res/layout/activity_main3.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple_200"
-    tools:context=".MainActivity3">
+    tools:context=".MainActivity">
 
     <TextView
         android:id="@+id/register"
@@ -142,4 +142,4 @@
         android:layout_marginEnd="20dp"
         android:layout_marginBottom="20dp"
         android:text="Sign Up" />
-
+</RelativeLayout>
